refactor: 인증 서비스 클래스 하나의 디렉토리로 관리
